refactor: Move android plugins to package android
